Issue #SC-1523 removed ununsed methods and created StateAdminReportHelper which contains common code for state and geo reports
